feat: ⚡️ start develop
- split files
- address bundle size, went from 12mb -> 1mb
- add support for selection sets (WIP)
