+++ .eslintrc.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   parser: "@typescript-eslint/parser", // Specifies the ESLint parser
   extends: [
     "eslint:recommended",
