+++ package.json
@@ -40,6 +40,10 @@
         "title": "Monorepo: Open Package (Current Window)"
       },
       {
+        "command": "mono-workspace.save_set",
+        "title": "Monorepo: Save Current Set"
+      },
+      {
         "command": "mono-workspace.openPackageNewWindow",
         "title": "Monorepo: Open Package (New Window)"
       },
@@ -63,6 +67,11 @@
           "type": "boolean",
           "default": true,
           "description": "Inlcude the top-level monorepo root path as a workspace folder"
+        },
+        "monoWorkspace.sets": {
+          "type": "array",
+          "default": [],
+          "description": "Predefined sets of workspaces. Usually defined in the monorepo root vscode setting file"
         },
         "monoWorkspace.providers_suffix": {
           "type": "boolean",
@@ -140,6 +149,7 @@
   "scripts": {
     "vscode:prepublish": "npm run compile",
     "compile": "node build.mjs",
+    "postcompile": "du -hs out/extension.js",
     "compile:dev": "node build.mjs",
     "lint": "eslint src --ext ts",
     "pretest": "npm run compile && npm run lint",
@@ -149,7 +159,6 @@
     "release": "vsce package --pre-release && vsce publish --pre-release"
   },
   "devDependencies": {
-    "@nrwl/devkit": "14.8.4",
     "@types/glob": "7.1.3",
     "@types/lodash": "^4.14.186",
     "@types/mocha": "8.2.1",
@@ -170,15 +179,18 @@
     "eslint-plugin-unicorn": "44.0.2",
     "glob": "7.1.6",
     "husky": "5.1.1",
-    "lodash": "^4.17.21",
     "mocha": "10.0.0",
     "prettier": "2.2.1",
-    "toml": "^3.0.0",
-    "tslib": "2.1.0",
     "typescript": "4.8.4",
-    "ultra-runner": "3.10.5",
     "vsce": "^2.11.0",
-    "vscode-test": "1.6.1",
-    "yaml": "^2.1.3"
+    "vscode-test": "1.6.1"
+  },
+  "dependencies": {
+    "jsonc": "^2.0.0",
+    "toml": "^3.0.0",
+    "lodash": "^4.17.21",
+    "@nrwl/devkit": "14.8.4",
+    "tslib": "2.1.0",
+    "ultra-runner": "3.10.5"
   }
 }
