+++ src/providers/core.ts
@@ -1,0 +1,6 @@
+import { getWorkspace } from "ultra-runner"
+import { GetProjectOptions } from "../types"
+
+export async function getMultiProjects(options: GetProjectOptions) {
+  return await getWorkspace(options)
+}
